requirements.txt
streamlit==1.37.0
google-generativeai==0.7.2
python-dotenv==1.0.1